Fix-details-route
