+++ src/App.jsx
@@ -7,7 +7,7 @@ import {
 import CartPage from "./pages/CartPage"
 import ErrorPage from "./pages/ErrorPage"
 import HomePage from "./pages/HomePage"
-import ProductDetailPage from "./components/product/ProductDetailPage"
+import ProductDetailPage from "./pages/ProductDetailPage"
 import Root from "./pages/Root"
 import TestPage from "./pages/TestPage"
 import ProductsPage from "./pages/ProductsPage"
